# openapi.yaml
openapi: 3.1.0
info:
  title: 6MD Management Platform API
  description: |
    Web application API for managing Arma 3 unit operations, signups, attendance, ranks, 
    training progression, and admin workflows for the 6th Maroon Division.
  version: 1.0.0
  contact:
    name: 6th Maroon Division
  license:
    name: Proprietary
    identifier: 6MD
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://orbat.6md.net/api
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /ranks/migrate/apply:
    post:
      tags:
        - Ranks
      summary: Apply rank migration
      security:
        - bearerAuth:
            - ranks:manage
      responses:
        "200":
          description: Migration applied
...
  /ranks/promotions/propose:
    post:
      tags:
        - Ranks
      summary: Propose promotion
      security:
        - bearerAuth:
            - ranks:propose_promotion
      responses:
        "201":
          description: Promotion proposed
...
  /signups/{id}/move:
    post:
      tags:
        - Signups
      summary: Move signup to different slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetSlotId:
                  type: integer
      responses:
        "200":
          description: Signup moved
        "401":
          description: Unauthorized
    patch:
      tags:
        - TODO
      summary: TODO document PATCH /signups/{id}/move
      responses:
        "200":
          description: Success
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth.js session token (cookie-based)
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: BotToken
      description: Active bot token stored in the BotToken database table. Use Authorization Bearer; environment tokens and X-BOT-API-TOKEN are not accepted.
...